Minor general refactor
