+++ src/MongoFramework/Infrastructure/Mapping/PropertyTraversalExtensions.cs
@@ -74,7 +74,7 @@ public static class PropertyTraversalExtensions
 				yield return traversedProperty;
 
 				var propertyType = traversedProperty.Property.PropertyInfo.PropertyType;
-				propertyType = propertyType.ElideEnumerableTypes();
+				propertyType = propertyType.UnwrapEnumerableTypes();
 
 				if (EntityMapping.IsValidTypeToMap(propertyType) && !state.SeenTypes.Contains(propertyType))
 				{
